fix: Reorganiza criação do tema principal do projeto
